TW-1711: Handle deep link in auto homeserver picker
